+++ ratarmount.py
@@ -471,7 +471,7 @@ class FuseMount(fuse.Operations):
 
         options['writeIndex'] = True
 
-        self.printDebug = options.get('printDebug', 0)
+        self.printDebug: int = options.get('printDebug', 0)
         self.writeOverlay: Optional[WritableFolderMountSource] = None
         self.overlayPath: Optional[str] = None
 
@@ -486,7 +486,7 @@ class FuseMount(fuse.Operations):
         mountSources = [openMountSource(path, **options) for path in pathToMount]
 
         self.mountSource: MountSource = UnionMountSource(mountSources, printDebug=self.printDebug)
-        if options.get('recursive', False):
+        if options.get('recursionDepth', 0):
             self.mountSource = AutoMountLayer(self.mountSource, **options)
 
         # No threads should be created and still be open before FUSE forks.
@@ -975,7 +975,20 @@ seeking capabilities when opening that file.
 
     parser.add_argument(
         '-r', '--recursive', action='store_true', default = False,
-        help = 'Mount TAR archives inside the mounted TAR recursively. '
+        help = 'Mount archives inside archives recursively. Same as --recursion-depth -1.' )
+
+    # TODO The recursion depth is only heeded by AutoMountLayer but not by SQLiteIndexedTar.
+    #      One problem is that it requires an update to the index metadata information and
+    #      the other problem is that the AutoMountLayer would have to ask how deep the recursion
+    #      for a particular path is so that it can correctly stop recursive mounting and the
+    #      combined recursion depth.
+    parser.add_argument(
+        '--recursion-depth', type = int,
+        help = 'This option takes precedence over --recursive. '
+               'Mount archives inside the mounted archives recursively up to the given depth. '
+               'A negative value represents infinite depth. '
+               'A value of 0 will turn off recursion (same as not specifying --recursive in the first place). '
+               'A value of 1 will recursively mount all archives in the given archives but not any deeper. '
                'Note that this only has an effect when creating an index. '
                'If an index already exists, then this option will be effectively ignored. '
                'Recreate the index if you want change the recursive mounting policy anyways.' )
@@ -1140,9 +1153,14 @@ seeking capabilities when opening that file.
 
     args.gzipSeekPointSpacing = args.gzip_seek_point_spacing * 1024 * 1024
 
-    if (args.strip_recursive_tar_extension or args.transform_recursive_mount_point) and not args.recursive:
+    if args.recursive and args.recursion_depth is None:
+        args.recursion_depth = -1
+    if args.recursion_depth is None:
+        args.recursion_depth = 0
+
+    if (args.strip_recursive_tar_extension or args.transform_recursive_mount_point) and not args.recursion_depth:
         print("[Warning] The options --strip-recursive-tar-extension and --transform-recursive-mount-point")
-        print("[Warning] only have an effect when used with --recursive.")
+        print("[Warning] only have an effect when used with recursive mounting.")
 
     if args.transform_recursive_mount_point:
         args.transform_recursive_mount_point = tuple(args.transform_recursive_mount_point)
@@ -1439,22 +1457,22 @@ def cli(rawArgs: Optional[List[str]] = None) -> None:
     fuseOperationsObject = FuseMount(
         # fmt: off
         pathToMount                  = args.mount_source,
-        clearIndexCache              = args.recreate_index,
-        recursive                    = args.recursive,
+        clearIndexCache              = bool(args.recreate_index),
+        recursionDepth               = int(args.recursion_depth),
         gzipSeekPointSpacing         = args.gzipSeekPointSpacing,
         mountPoint                   = args.mount_point,
         encoding                     = args.encoding,
-        ignoreZeros                  = args.ignore_zeros,
-        verifyModificationTime       = args.verify_mtime,
+        ignoreZeros                  = bool(args.ignore_zeros),
+        verifyModificationTime       = bool(args.verify_mtime),
         stripRecursiveTarExtension   = args.strip_recursive_tar_extension,
         indexFilePath                = args.index_file,
         indexFolders                 = args.index_folders,
-        lazyMounting                 = args.lazy,
-        passwords                    = args.passwords,
+        lazyMounting                 = bool(args.lazy),
+        passwords                    = list(args.passwords),
         parallelization              = args.parallelization,
         isGnuIncremental             = args.gnu_incremental,
         writeOverlay                 = args.write_overlay,
-        printDebug                   = args.debug,
+        printDebug                   = int(args.debug),
         transformRecursiveMountPoint = args.transform_recursive_mount_point,
         # fmt: on
     )
@@ -1470,7 +1488,7 @@ def cli(rawArgs: Optional[List[str]] = None) -> None:
             **fusekwargs
         )
     except RuntimeError as exception:
-        raise RuntimeError(
+        raise RatarmountError(
             "FUSE mountpoint could not be created. See previous output for more information."
         ) from exception